refactor(service/controller): realoca controllers e serviços por domínio
