+++ src/main/java/com/project/miinhareceita/favorite/controller/FavoriteController.java
@@ -1,8 +1,8 @@
-package com.project.miinhareceita.controllers;
+package com.project.miinhareceita.favorite.controller;
 
 import com.project.miinhareceita.favorite.dto.FavoriteDTO;
 import com.project.miinhareceita.favorite.dto.FavoriteInsertDTO;
-import com.project.miinhareceita.services.FavoriteService;
+import com.project.miinhareceita.favorite.service.FavoriteService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
