+++ src/main/java/com/project/miinhareceita/recipe/service/RecipeService.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.services;
+package com.project.miinhareceita.recipe.service;
 
 import com.project.miinhareceita.auth.service.AuthService;
 import com.project.miinhareceita.ingredient.domain.Ingredients;
